only authenticated user can access and manipulate his own goals

+++ backend/controllers/goalController.js
@@ -1,11 +1,11 @@
 const asyncHandler = require("express-async-handler");
-
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 //  @desc    get goals
 //  @route   GET /api/goals
 //  @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -20,6 +20,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -36,6 +37,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error("Goal not found");
   }
 
+  const user = await User.findById(req.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //  Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -47,18 +62,31 @@ const updateGoal = asyncHandler(async (req, res) => {
 // @route   DELETE /api/goals/:id
 // @access  Private
 const deleteGoal = asyncHandler(async (req, res) => {
-  const goal = await Goal.findById(req.params.id)
+  const goal = await Goal.findById(req.params.id);
 
   if (!goal) {
-    res.status(400)
-    throw new Error('Goal not found')
+    res.status(400);
+    throw new Error("Goal not found");
   }
 
-  await goal.deleteOne()
+  const user = await User.findById(req.user.id);
 
-  res.status(200).json({ id: req.params.id })
-})
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
 
+  //  Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
+  }
+
+  await goal.deleteOne();
+
+  res.status(200).json({ id: req.params.id });
+});
 
 module.exports = {
   getGoals,
